+++ src/domains/scrobbleSetlist/ScrobbleSetlistView.tsx
@@ -3,7 +3,6 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { ClearHistoryButton } from 'components/ClearHistoryButton';
 import EmptyScrobbleListFiller from 'components/EmptyScrobbleListFiller';
 import ScrobbleList from 'components/ScrobbleList';
-import Spinner from 'components/Spinner';
 import SetlistViewer from 'domains/scrobbleSetlist/Setlist';
 import { Trans } from 'react-i18next';
 import { useSelector } from 'react-redux';
